Added CartridgeInspector

+++ Emulator/VICII/VICII.cpp
@@ -553,7 +553,7 @@ VICII::clearStats()
 }
 
 SpriteInfo
-VICII::getSpriteInfo(int nr)
+VICII::getSpriteInfo(isize nr)
 {
     {   SYNCHRONIZED
